// frontend/src/lib/generated-bindings.ts
// This file has been generated by Specta. Do not edit this file manually.
export type Stage = {
	id: number;
	title: string;
	status: TaskStatus;
	tasks: Task[];
};

export type SubTask = {
	id: number;
	parent: number;
	title: string;
	status: TaskStatus;
	description: string;
};

export type Task = {
	id: number;
	parent: number;
	title: string;
	status: TaskStatus;
	description: string;
	subtasks: SubTask[];
};

export type TaskStatus = "pending" | "in-progress" | "done";

export type WsEvent =
	| ({ AddTask: Task } & {
			AddStage?: never;
			AddSubTask?: never;
			Delete?: never;
			UpdateStage?: never;
			UpdateSubTask?: never;
			UpdateTask?: never;
	  })
	| ({ UpdateTask: Task } & {
			AddStage?: never;
			AddSubTask?: never;
			AddTask?: never;
			Delete?: never;
			UpdateStage?: never;
			UpdateSubTask?: never;
	  })
	| ({ AddSubTask: SubTask } & {
			AddStage?: never;
			AddTask?: never;
			Delete?: never;
			UpdateStage?: never;
			UpdateSubTask?: never;
			UpdateTask?: never;
	  })
	| ({ UpdateSubTask: SubTask } & {
			AddStage?: never;
			AddSubTask?: never;
			AddTask?: never;
			Delete?: never;
			UpdateStage?: never;
			UpdateTask?: never;
	  })
	| ({ AddStage: Stage } & {
			AddSubTask?: never;
			AddTask?: never;
			Delete?: never;
			UpdateStage?: never;
			UpdateSubTask?: never;
			UpdateTask?: never;
	  })
	| ({ UpdateStage: Stage } & {
			AddStage?: never;
			AddSubTask?: never;
			AddTask?: never;
			Delete?: never;
			UpdateSubTask?: never;
			UpdateTask?: never;
	  })
	| ({ Delete: number } & {
			AddStage?: never;
			AddSubTask?: never;
			AddTask?: never;
			UpdateStage?: never;
			UpdateSubTask?: never;
			UpdateTask?: never;
	  });